Implemented proper periodic boundary conditions for Cubic bezier rings.
The algorithm is from

  http://www.codeproject.com/KB/graphics/ClosedBezierSpline.aspx

However, the page contains an error which messes up the result. The
indexing of the second control point is off by one, and needs to
be periodic with respect to the original knot points. The problematic
part is marked in the code, and the original version is in an #else
section which is preprocessed out.