Added jet rendering.
There seems to be a bug in Chrome regarding the rendering of end-markers.
If the tip of the marker is extended to0 far of the end of the spline,
Chrome will not extend the path and instead will generate an extra knot off
to the side of the tip, making the arrow unsymmetric. The tip is now
as far as I could extend it.