Added front letter spacing and graphic symbol customization.
The size of SVG symbols cannot be customized using CSS or SVG defs.
As an extension frontier is now able to find the following setting
from the CSS section:

 .GraphicSymbol
 {
   fmi-size: 59px;
 }

Also, the spacing of warm front, cold front, occluded front, trough
and upper trough symbols can now be controlled as follows using
weather object specific settings:

 .coldfrontglyph
 {
   fmi-letter-spacing: 60px;
 }

In both cases the units must be in pixels.

Note that the latter setting should not be required, the normal CSS
letter-spacing should work, but the currently available programs
do not support it properly. In particular, they may start the
letters at offset 0 even though a specific offset is given,
and the behaviour may depend on whether there is an even or
odd number of characters in the string being shown. Once the programs
support the attribute properly, we may simply drop the "fmi-" prefix
from the name.