Fixed alignment issues.
For numbers we need to use "alignment-baseline: middle" to get the numbers
centered the same way as in Mirwa.

For graphic symbols we cannot use style sheets, so we need to substract
half the size of the symbol from x- and y-coordinates to get the symbol
to appear in the same place as in Mirwa.