Separated styling of point values and ranges.
For example, if there is a point value and a range for a parameter
named Temperature, the respective CSS style is no longer Temperature
but Temperaturevalue or Temperaturerange depending on the type
of the symbol. The separation is associated with the changes required
for supporting contour styles (Temperaturecontour), simply using the
parameter name for CSS classes is no longer sufficient.