Added code for reading a libconfig configuration from the SVG template.
However, RHEL currently provides libconfig-1.3-1, and the readString
method we need is not implemented in it. The latest libconfig release
version is 1.4.5. We'll need to make a fmiforge package to provide
a newer release, or write the libconfig string temporarily to a
file. Neither choice sounds good. Since libconfig is dynamically
linked being an officially supported package, we better not mess
with the RHEL distribution by introducing a newer libconfig which
might break some package in the official RHEL distribution.
This leaves no other choice than to write the extracted config to a
temp file, and read it back using libconfig::Config.read(), which
I'll implement next.