Added extraction of <frontier> libconfig part from the template.
The code now has to extract the configuration, write it to a file,
read it with libconfig++ readFile, and unlink the temporary file
because our libconfig++ version is too old. Newer versions support
a readString method. The code now contains and #if portion
to identify the spots which can be simplified when a new libconfig++
is available.