Fixed libconfig++ usage.
libconfig provides poor error messages when it fails to parse
values. As a consequence I wasted a lot of time trying to understand
why

 occludedfront:
 {
     letter-spacing = 60;
 };

does not parse. The reason was that I was assigning to a double,
and there is no decimal point in that number.

There is now a wrapper lookup function which informs:

  a) if the setting is missing completely
  b) the name of the variable being looked up
  c) the desired type of the variable