Intermediate commit which reveals a problem libwoml API.
According to boost specs a period defined with starttime==endtime
is invalid. In fact,

 a) if you print the time period, the end time will be a microsecond
    smaller than the starttime, regardless of your input.

 b) invalid periods are not considered equal even if the start and
    end times are equal.

Hence one cannot easily collect a set of unique time periods.
According to gml, a valid time may be a time instant or a time period,
and since we cannot represent a time instant as a time period of
length 0, libwoml API has to be extended.